infrastructure/channels/test_executeScript.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>TestDriver execute method (async)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/channel.sub.js"></script>

<script>
setup({single_test: true})
onload = async () => {
    let remote = await new RemoteWindow();

    let url = `child_slow.html?uuid=${remote.uuid}`;
    win = window.open(url, "_blank", "noopener");

    let result = await remote.executeScript(async (elemId) => {
        await new Promise((resolve) => setTimeout(resolve, 1000));
        return document.getElementById(elemId).textContent;
    }, ["test"]);
    assert_equals(result.trim(), "PASS");
    done();
}
</script>
